pyproject.toml
[project]
name = "e-commerce-app"
version = "0.1.0"
description = "A simple e commerce api"
authors = [
    {name = "Derrick",email = "derrick.sarpong.ds@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "pydantic (>=2.10.6,<2.11.0)",
    "sqlalchemy (>=2.0.38,<2.1.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<1.1.0)",
    "pyicu (>=2.14,<3.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "falcon (>=4.0.2,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
